+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/environment/Vocabulary.java
@@ -39,6 +39,8 @@ public class Vocabulary {
     public static final String c_OwlClassS = CLASS_BASE + "OWLClassS";
     public static final String c_OwlClassT = CLASS_BASE + "OWLClassT";
     public static final String c_OwlClassU = CLASS_BASE + "OWLClassU";
+    public static final String c_OwlClassV = CLASS_BASE + "OWLClassV";
+    public static final String c_OwlClassW = CLASS_BASE + "OWLClassW";
     public static final String c_OwlClassWithQueryAttr = CLASS_BASE + "OWLClassWithQueryAttr";
     public static final String c_Person = CLASS_BASE + "Person";
     public static final String c_Phone = CLASS_BASE + "Phone";
@@ -81,6 +83,13 @@ public class Vocabulary {
 
     public static final String P_U_SINGULAR_MULTILINGUAL_ATTRIBUTE = ATTRIBUTE_BASE + "singularMultilingualAttribute";
     public static final String P_U_PLURAL_MULTILINGUAL_ATTRIBUTE = ATTRIBUTE_BASE + "pluralMultilingualAttribute";
+
+    public static final String P_V_SINGULAR_DYNAMIC_ATTRIBUTE = ATTRIBUTE_BASE + "singularDynamicAttribute";
+    public static final String P_V_PLURAL_DYNAMIC_ATTRIBUTE = ATTRIBUTE_BASE + "pluralDynamicAttribute";
+
+    public static final String P_W_SET_STRING_ATTRIBUTE = ATTRIBUTE_BASE + "setStringAttribute";
+    public static final String P_W_LIST_STRING_ATTRIBUTE = ATTRIBUTE_BASE + "listStringAttribute";
+    public static final String P_W_COLLECTION_STRING_ATTRIBUTE = ATTRIBUTE_BASE + "collectionStringAttribute";
 
     public static final String p_p_username = ATTRIBUTE_BASE + "username";
     public static final String p_p_age = ATTRIBUTE_BASE + "age";