+++ jopa-integration-tests/src/main/java/cz/cvut/kbss/jopa/test/runner/RetrieveOperationsRunner.java
@@ -23,6 +23,7 @@ import cz.cvut.kbss.jopa.model.descriptors.EntityDescriptor;
 import cz.cvut.kbss.jopa.model.query.TypedQuery;
 import cz.cvut.kbss.jopa.proxy.lazy.LazyLoadingProxy;
 import cz.cvut.kbss.jopa.test.OWLClassA;
+import cz.cvut.kbss.jopa.test.OWLClassAA;
 import cz.cvut.kbss.jopa.test.OWLClassB;
 import cz.cvut.kbss.jopa.test.OWLClassC;
 import cz.cvut.kbss.jopa.test.OWLClassD;
@@ -36,6 +37,7 @@ import cz.cvut.kbss.jopa.test.OWLClassV;
 import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr;
 import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr2;
 import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr6;
+import cz.cvut.kbss.jopa.test.OWLClassWithQueryAttr7;
 import cz.cvut.kbss.jopa.test.OWLClassWithUrn;
 import cz.cvut.kbss.jopa.test.Thing;
 import cz.cvut.kbss.jopa.test.Vocabulary;
@@ -61,6 +63,7 @@ import java.nio.file.StandardCopyOption;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -542,6 +545,45 @@ public abstract class RetrieveOperationsRunner extends BaseRunner {
         assertEquals(entityD.getUri(), res.getLazyQueryAttribute().getUri());
     }
 
+
+    @Test
+    public void testRetrieveWithCollectionQueryAttribute() {
+        this.em = getEntityManager("RetrieveWithPluralQueryAttribute", false);
+
+        persist(entityWithQueryAttr7);
+        final OWLClassWithQueryAttr7 res = findRequired(OWLClassWithQueryAttr7.class, entityWithQueryAttr7.getUri());
+        Collection<Integer> result = res.getCollectionQueryAttribute();
+        assertInstanceOf(List.class, result);
+        assertEquals(3, result.size());
+        assertEquals(2, result.stream().filter(i -> i.equals(99)).count());
+    }
+
+    @Test
+    public void testRetrieveWithSetQueryAttribute() {
+        this.em = getEntityManager("RetrieveWithPluralQueryAttribute", false);
+
+        persist(entityWithQueryAttr7);
+        final OWLClassWithQueryAttr7 res = findRequired(OWLClassWithQueryAttr7.class, entityWithQueryAttr7.getUri());
+        Set<Integer> result = res.getSetQueryAttribute();
+        assertInstanceOf(Set.class, result);
+        assertEquals(2, result.size());
+        assertTrue(result.containsAll(List.of(99, 2)));
+    }
+
+    @Test
+    public void testRetrieveWithListQueryAttribute() {
+        this.em = getEntityManager("RetrieveWithPluralQueryAttribute", false);
+
+        persist(entityWithQueryAttr7);
+        final OWLClassWithQueryAttr7 res = findRequired(OWLClassWithQueryAttr7.class, entityWithQueryAttr7.getUri());
+        Collection<Integer> result = res.getCollectionQueryAttribute();
+        assertInstanceOf(List.class, result);
+        assertEquals(3, result.size());
+        assertEquals(2, result.stream().filter(i -> i.equals(99)).count());
+    }
+
+
+
     @Test
     void testSupportForUrnIrisInClassAndProperty() {
         this.em = getEntityManager("testSupportForUrnIrisInClassAndProperty", false);
@@ -609,5 +651,35 @@ public abstract class RetrieveOperationsRunner extends BaseRunner {
 
         final OWLClassI result = findRequired(OWLClassI.class, entityI.getUri());
         assertNull(result.getOwlClassA());
+    }
+
+    @Test
+    public void retrieveDynamicStringAttribute() {
+        this.em = getEntityManager("retrieveDynamicAttribute", false);
+        entityAA.setDynamicProperty("Hello, world!");
+        transactional(() -> em.persist(entityAA));
+
+        final OWLClassAA result = findRequired(OWLClassAA.class, entityAA.getUri());
+        assertEquals("Hello, world!", result.getDynamicProperty());
+    }
+
+    @Test
+    public void retrieveDynamicIntAttribute() {
+        this.em = getEntityManager("retrieveDynamicAttribute", false);
+        entityAA.setDynamicProperty(1234);
+        transactional(() -> em.persist(entityAA));
+
+        final OWLClassAA result = findRequired(OWLClassAA.class, entityAA.getUri());
+        assertEquals(1234, result.getDynamicProperty());
+    }
+
+    @Test
+    public void retrieveDynamicDoubleAttribute() {
+        this.em = getEntityManager("retrieveDynamicAttribute", false);
+        entityAA.setDynamicProperty(Double.parseDouble("1234.6970"));
+        transactional(() -> em.persist(entityAA));
+
+        final OWLClassAA result = findRequired(OWLClassAA.class, entityAA.getUri());
+        assertEquals(1234.697D, result.getDynamicProperty());
     }
 }